Added support for the field name type.
Since type is a keyword in Julia we need to make sure that if the protobuf
message has a field name 'type' that we do not use that for the Julia's type
attribute name. Instead we'll replace it with '_type'.

This same solution or something similar should likely be employed for other
Julia keywords.