Preserve comments in output _pb.jl files from original .proto
This commit adds code to `generate_file` to copy all the comments from
the original `.proto` file into the new generated `_pb.jl` file by
reading the comments fields off of the SourceCodeInfo proto.

The comments mechanism is a bit confusing in the proto descriptors, so I
separated out most of the newly added logic into a new file,
`src/commetns.jl`.

The changes in this commit were inspired by following the example of the
code here (which is under the MIT license):
https://github.com/pseudomuto/protokit/blob/7037620/comments.go